requirements.txt
langchain
langchain-community
langchain-ollama
langchain-qdrant
langchain-openai
langgraph
streamlit
qdrant-client
pypdf
docx2txt
tiktoken
unstructured
deepeval
pytest
arize-phoenix
openinference-instrumentation-langchain
flashrank